script: Move ElementInternals into `script/dom/html/internals/`
I also wanted to remove the final usage from `Element`, but currently
it is on `rare_data()`, which means that it is private to `Element`.
Maybe we should in the future create a `HTMLElementRareData`?

Testing: it compiles
