+++ components/script/dom/mod.rs
@@ -236,7 +236,6 @@ pub(crate) use self::credentialmanagement::*;
 pub(crate) mod css;
 pub(crate) use self::css::*;
 pub(crate) mod customelementregistry;
-pub(crate) mod customstateset;
 pub(crate) mod datatransfer;
 pub(crate) use self::datatransfer::*;
 pub(crate) mod debugger;
@@ -251,7 +250,6 @@ pub(crate) mod domstringmap;
 pub(crate) mod domtokenlist;
 pub(crate) mod element;
 pub(crate) use self::element::*;
-pub(crate) mod elementinternals;
 pub(crate) mod encoding;
 pub(crate) use self::encoding::*;
 pub(crate) mod event;
